requirements.txt
Flask==3.1.0
gunicorn==21.2.0
gdown==5.1.0

tensorflow==2.17.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
keras==3.5.0
h5py==3.13.0
protobuf==4.25.6
gast==0.4.0
opt_einsum==3.4.0
ml-dtypes==0.3.2
wrapt==1.14.1
dm-tree==0.1.9

jax==0.4.30
jaxlib==0.4.30
optree==0.15.0

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.15.2
joblib==1.4.2
threadpoolctl==3.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

pillow==11.2.1

requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

google-auth==2.39.0
google-auth-oauthlib==1.0.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
rsa==4.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
google-pasta==0.2.0

absl-py==2.2.2
libclang==18.1.1
Markdown==3.8
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1

markdown-it-py==3.0.0
mdurl==0.1.2
rich==14.0.0
Pygments==2.19.1
attrs==25.3.0

astunparse==1.6.3
click==8.1.8
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.13.2
